+++ src/components/common/Modal/AddressModal.tsx
@@ -1,14 +1,14 @@
 import { Dispatch, ReactNode } from 'react';
 
 import Icon from '@/components/common/Icon';
-import { ModalContainer, ModalXBtnBox } from '@/components/common/Modal/Modal.styles';
+import { ModalContainer, ModalXBtnBox } from '@/components/common/Modal/AddressModal.styles';
 
 interface ModalProps {
   children: ReactNode;
   setModalOpen: Dispatch<React.SetStateAction<boolean>>;
 }
 
-const Modal = ({ children, setModalOpen }: ModalProps) => {
+const AddressModal = ({ children, setModalOpen }: ModalProps) => {
   const handleCloseModal = () => {
     setModalOpen(false);
   };
@@ -22,4 +22,4 @@ const Modal = ({ children, setModalOpen }: ModalProps) => {
     </ModalContainer>
   );
 };
-export default Modal;
+export default AddressModal;
